Update : cmpoound component pattern 대신 기본 컴포넌트 사용
